UI에 현재 콥스수, 맥스, 리젠타임 보여줌 및 클라 텍스트 렌더수정
TextManager::GetInstance()->DrawTexts();
-클라 렌더러에 그냥 싱글톤 호출했는데 검토좀
-리젠 공식 너무 타임이 안 올라감 수정 필요

+++ YaMang/YaMangServer/YaMangServer/GameRoom.cpp
@@ -439,7 +439,7 @@ void GameRoom::TakeBase( int ownerPlayerID, int targetPlayerID, int ownerCorpsID
 		return;
 	}
 
-	int targetGuardListID = std::find( m_GuardIDList.begin( ), m_GuardIDList.end( ), targetGuardID ) - m_GuardIDList.begin( );
+	int targetGuardListID = static_cast<int> ( std::find( m_GuardIDList.begin(), m_GuardIDList.end(), targetGuardID ) - m_GuardIDList.begin() );
 	if ( -1 == targetGuardListID )
 	{
 		return;
